productivity/dev_productivity_app/requirements.txt
# Production dependencies
matplotlib>=3.8.0
numpy>=1.26.0
jsonschema>=4.20.0
pyyaml>=6.0.0

# Development/Testing dependencies
pytest>=8.0.0
pytest-cov>=4.1.0
pytest-mock>=3.12.0
